Extract option component
